+++ common/src/comp/ability.rs
@@ -39,7 +39,7 @@ impl From<&CharacterState> for CharacterAbilityType {
             CharacterState::BasicRanged(_) => Self::BasicRanged,
             CharacterState::Boost(_) => Self::Boost,
             CharacterState::DashMelee(data) => Self::DashMelee(data.stage_section),
-            CharacterState::BasicBlock => Self::BasicBlock,
+            CharacterState::BasicBlock(_) => Self::BasicBlock,
             CharacterState::LeapMelee(data) => Self::LeapMelee(data.stage_section),
             CharacterState::ComboMelee(data) => Self::ComboMelee(data.stage_section, data.stage),
             CharacterState::SpinMelee(data) => Self::SpinMelee(data.stage_section),
@@ -114,7 +114,13 @@ pub enum CharacterAbility {
         charge_through: bool,
         is_interruptible: bool,
     },
-    BasicBlock,
+    BasicBlock {
+        buildup_duration: f32,
+        recover_duration: f32,
+        max_angle: f32,
+        block_strength: f32,
+        energy_cost: f32,
+    },
     Roll {
         energy_cost: f32,
         buildup_duration: f32,
@@ -305,7 +311,8 @@ impl CharacterAbility {
             | CharacterAbility::ChargedRanged { energy_cost, .. }
             | CharacterAbility::ChargedMelee { energy_cost, .. }
             | CharacterAbility::Shockwave { energy_cost, .. }
-            | CharacterAbility::BasicAura { energy_cost, .. } => update
+            | CharacterAbility::BasicAura { energy_cost, .. }
+            | CharacterAbility::BasicBlock { energy_cost, .. } => update
                 .energy
                 .try_change_by(-(*energy_cost as i32), EnergySource::Ability)
                 .is_ok(),
@@ -338,6 +345,16 @@ impl CharacterAbility {
             recover_duration: 0.125,
             roll_strength: 2.0,
             immune_melee: true,
+        }
+    }
+
+    pub fn default_block() -> CharacterAbility {
+        CharacterAbility::BasicBlock {
+            buildup_duration: 0.35,
+            recover_duration: 0.3,
+            max_angle: 60.0,
+            block_strength: 0.5,
+            energy_cost: 50.0,
         }
     }
 
@@ -408,7 +425,15 @@ impl CharacterAbility {
                 *swing_duration /= speed;
                 *recover_duration /= speed;
             },
-            BasicBlock => {},
+            BasicBlock {
+                ref mut buildup_duration,
+                ref mut recover_duration,
+                // Block strength explicitly not modified by power, that will be a separate stat
+                ..
+            } => {
+                *buildup_duration /= speed;
+                *recover_duration /= speed;
+            },
             Roll {
                 ref mut buildup_duration,
                 ref mut movement_duration,
@@ -578,7 +603,8 @@ impl CharacterAbility {
             | ChargedRanged { energy_cost, .. }
             | Shockwave { energy_cost, .. }
             | HealingBeam { energy_cost, .. }
-            | BasicAura { energy_cost, .. } => *energy_cost as u32,
+            | BasicAura { energy_cost, .. }
+            | BasicBlock { energy_cost, .. } => *energy_cost as u32,
             BasicBeam { energy_drain, .. } => {
                 if *energy_drain > f32::EPSILON {
                     1
@@ -586,7 +612,7 @@ impl CharacterAbility {
                     0
                 }
             },
-            BasicBlock | Boost { .. } | ComboMelee { .. } | Blink { .. } | BasicSummon { .. } => 0,
+            Boost { .. } | ComboMelee { .. } | Blink { .. } | BasicSummon { .. } => 0,
         }
     }
 
@@ -1235,7 +1261,23 @@ impl From<(&CharacterAbility, AbilityInfo)> for CharacterState {
                 stage_section: StageSection::Buildup,
                 exhausted: false,
             }),
-            CharacterAbility::BasicBlock => CharacterState::BasicBlock,
+            CharacterAbility::BasicBlock {
+                buildup_duration,
+                recover_duration,
+                max_angle,
+                block_strength,
+                energy_cost: _,
+            } => CharacterState::BasicBlock(basic_block::Data {
+                static_data: basic_block::StaticData {
+                    buildup_duration: Duration::from_secs_f32(*buildup_duration),
+                    recover_duration: Duration::from_secs_f32(*recover_duration),
+                    max_angle: *max_angle,
+                    block_strength: *block_strength,
+                    ability_info,
+                },
+                timer: Duration::default(),
+                stage_section: StageSection::Buildup,
+            }),
             CharacterAbility::Roll {
                 energy_cost: _,
                 buildup_duration,