+++ common/src/comp/controller.rs
@@ -149,15 +149,19 @@ impl ControlAction {
 pub enum InputKind {
     Primary = 0,
     Secondary = 1,
-    Ability(usize) = 2,
-    Roll = 3,
-    Jump = 4,
-    Fly = 5,
+    Block = 2,
+    Ability(usize) = 3,
+    Roll = 4,
+    Jump = 5,
+    Fly = 6,
 }
 
 impl InputKind {
     pub fn is_ability(self) -> bool {
-        matches!(self, Self::Primary | Self::Secondary | Self::Ability(_))
+        matches!(
+            self,
+            Self::Primary | Self::Secondary | Self::Ability(_) | Self::Block
+        )
     }
 }