+++ common/src/outcome.rs
@@ -59,6 +59,11 @@ pub enum Outcome {
     Damage {
         pos: Vec3<f32>,
     },
+    Block {
+        pos: Vec3<f32>,
+        parry: bool,
+        uid: Uid,
+    },
 }
 
 impl Outcome {
@@ -70,7 +75,8 @@ impl Outcome {
             | Outcome::Beam { pos, .. }
             | Outcome::SkillPointGain { pos, .. }
             | Outcome::SummonedCreature { pos, .. }
-            | Outcome::Damage { pos, .. } => Some(*pos),
+            | Outcome::Damage { pos, .. }
+            | Outcome::Block { pos, .. } => Some(*pos),
             Outcome::BreakBlock { pos, .. } => Some(pos.map(|e| e as f32 + 0.5)),
             Outcome::ExpChange { .. } | Outcome::ComboChange { .. } => None,
         }